refactor: Result 클래스 제거
