+++ src/main/java/racingcar/model/NumberGenerator.java
@@ -1,4 +1,4 @@
-package racingcar;
+package racingcar.model;
 
 public interface NumberGenerator {
